requirements.txt
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
ark-nlp==0.0.7
attrs==21.4.0
backcall==0.2.0
bcrypt==3.2.0
bigquery==0.0.20
bleach==4.1.0
boto3==1.21.7
botocore==1.24.7
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.4
colorama==0.4.6
colorlog==6.7.0
cryptography==36.0.1
cycler==0.11.0
dacktool==0.0.7
dbstream==0.1.13
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
filelock==3.6.0
fonttools==4.34.4
gensim==4.1.2
gitdb==4.0.10
GitPython==3.1.30
google-api-core==2.3.2
google-api-python-client==1.7.11
google-auth==2.6.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.6
google-cloud-bigquery==2.32.0
google-cloud-bigquery-storage==2.11.0
google-cloud-core==2.2.2
google-cloud-secret-manager==2.7.2
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
googleauthentication==0.0.16
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
grpcio-status==1.43.0
hmmlearn==0.2.8
httplib2==0.20.4
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.0
importlib-resources==5.4.0
ipdb==0.13.9
ipykernel==6.9.0
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jieba==0.42.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaggle==1.5.12
kiwisolver==1.4.3
libcst==0.4.1
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mpmath==1.2.1
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.7
notebook==6.4.8
numpy==1.21.5
oauthlib==3.2.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
paramiko==2.9.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prettytable==3.2.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
proto-plus==1.20.0
protobuf==3.19.4
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyBKT==1.4
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.13.0
seqeval==1.2.2
setproctitle==1.3.2
SimpleHMM==1.0.0
six==1.16.0
sklearn==0.0
sklearn-crfsuite==0.3.6
smart-open==5.2.1
smmap==5.0.0
sshtunnel==0.1.5
sympy==1.10.1
tabulate==0.9.0
tensorboardX==2.5.1
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.1.0
tokenizers==0.11.6
toml==0.10.2
torch==1.9.0+cu111
torchaudio==0.9.0
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.17.0
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.14
wandb==0.13.9
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
yata==0.1
zhon==1.1.5
zipp==3.7.0
